+++ sync-for-expenses/src/codatsyncexpenses/models/shared/contactref.py
@@ -4,19 +4,14 @@ from __future__ import annotations
 import dataclasses
 from codatsyncexpenses import utils
 from dataclasses_json import Undefined, dataclass_json
-from enum import Enum
-from typing import Optional
-
-class ContactRefContactType(str, Enum):
-    r"""The type of contact."""
-    SUPPLIER = 'Supplier'
+from typing import Final, Optional
 
 
 @dataclass_json(undefined=Undefined.EXCLUDE)
 
 @dataclasses.dataclass
 class ContactRef:
-    contact_type: Optional[ContactRefContactType] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('contactType'), 'exclude': lambda f: f is None }})
+    CONTACT_TYPE: Final[Optional[str]] = dataclasses.field(default='Supplier', metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('contactType'), 'exclude': lambda f: f is None }})
     r"""The type of contact."""
     id: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('id'), 'exclude': lambda f: f is None }})
     r"""Identifier of supplier or customer."""