+++ sync-for-expenses/src/codatsyncexpenses/sdkconfiguration.py
@@ -19,8 +19,9 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = 'prealpha'
-    sdk_version: str = '2.2.1'
-    gen_version: str = '2.131.1'
+    sdk_version: str = '2.3.0'
+    gen_version: str = '2.139.1'
+    user_agent: str = 'speakeasy-sdk/python 2.3.0 2.139.1 prealpha codat-sync-for-expenses'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> tuple[str, dict[str, str]]: